Assignment/Personal Library Manager/Personal-Library-Manager/pyproject.toml
[project]
name = "personal-library-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "panda>=0.3.1",
    "pillow>=11.2.1",
    "streamlit>=1.45.1",
    "watchdog>=6.0.0",
]
